**Refactor namespace structure and update Git version placeholders across throwable and support classes**
- Moved exceptions to a more granular namespace structure for clarity:
  - `InvalidStringEncodingError` and `CodepointOutsideValidRangeError` relocated from `FireHub\Core\Throwable\Error\LowLevel` to `FireHub\Core\Throwable\Error\LowLevel\String`.
  - Updated affected imports and docblocks in `CharMB`, `CharSB`, and other impacted classes to reflect the new namespace.

- Updated Git version placeholders (`$Id`) in docblocks:
  - Replaced specific checksums with standardized `$Id$` placeholders across all modified files to maintain consistency and simplify version tracking.

These changes improve maintainability by introducing a more intuitive hierarchy for throwable classes, while standardizing versioning for easier future updates.
